refactor: 메인 페이지 변경
