--- src/pages/show.tsx
@@ -1,38 +1,0 @@
-import Head from 'next/head';
-import styled from '@emotion/styled';
-import { Grid } from '@chakra-ui/react';
-
-import { DUMMY_MOVIE } from '@/constants/dummy';
-import { Bottom, LinedTitle, MovieCard } from '@/components';
-
-export default function ShowPage() {
-  return (
-    <>
-      <Head>
-        <title>Create Next Apps</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Content>
-        <LinedTitle title={'현재 상영작'} />
-        <Grid templateColumns="repeat(4, 1fr)" rowGap={24} columnGap={8} mt={8}>
-          {[...Array(12)].fill(DUMMY_MOVIE).map((movie, idx) => (
-            <MovieCard movie={movie} key={idx} />
-          ))}
-        </Grid>
-      </Content>
-      <Bottom />
-    </>
-  );
-}
-
-const Content = styled.div`
-  display: block;
-  padding-inline: 40px;
-
-  ${({ theme }) =>
-    `@media (max-width: ${theme.bp.mobile}) {
-      padding-inline: 24px;
-    }`}
-`;
